samples/python/agents/helloworld_with_ratelimiter/pyproject.toml
[project]
name = "helloworld-with-ratelimiter"
version = "0.1.0"
description = "HelloWorld agent example demonstrating rate limiting extension"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.5",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]

[project.optional-dependencies]
ratelimiter = [
    # Local development - would be "ratelimiter-ext" in production
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
